Cleanup parser namespace
