+++ lib/opal/cli.rb
@@ -114,7 +114,7 @@ module Opal
 
     def show_sexp
       evals_or_file do |contents, filename|
-        buffer = ::Opal::Source::Buffer.new(filename)
+        buffer = ::Opal::Parser::SourceBuffer.new(filename)
         buffer.source = contents
         sexp = Opal::Parser.default_parser.parse(buffer)
         output.puts sexp.inspect
