+++ lib/opal/parser/with_c_lexer.rb
@@ -3,11 +3,11 @@
 begin
   require 'c_lexer'
 rescue LoadError
-  $stderr.puts 'Failed to load CLexer, using pure Ruby lexer'
+  $stderr.puts 'Failed to load WithCLexer, using pure Ruby lexer'
 end
 
-if defined? Parser::Ruby25WithCLexer
-  class Opal::Parser::WithCLexer < Parser::Ruby25WithCLexer
+if defined? Parser::Ruby25WithWithCLexer
+  class Opal::Parser::WithCLexer < Parser::Ruby25WithWithCLexer
     include Opal::Parser::DefaultConfig
     Opal::Parser.default_parser_class = self
   end
